renamed model to registration for easier viewing and less insanity
